+++ src/api/server.ts
@@ -120,6 +120,7 @@ const getTransactionsByAccount = require('./earmark/getTransactionsByAccount')
 const allTransactionsByCategory = require('./earmark/allTransactionsByCategory');
 const recurring = require('./earmark/recurring');
 const earmark_dashboard = require('./earmark/dashboard');
+const earmark_visuals = require('./earmark/visuals');
 
 app.use('/api/earmark/allAccountInfo', validateApiKey, earmark_allAccountInfo);
 app.use('/api/earmark/allTransactions', validateApiKey, earmark_allTransactions);
@@ -129,7 +130,7 @@ app.use('/api/earmark/getTransactionsByAccount', validateApiKey, getTransactions
 app.use('/api/earmark/allTransactionsByCategory', validateApiKey, allTransactionsByCategory);
 app.use('/api/earmark/recurring', validateApiKey, recurring);
 app.use('/api/earmark/dashboard', validateApiKey, earmark_dashboard);
-
+app.use('/api/earmark/visuals', validateApiKey, earmark_visuals);
 // firebase routes
 const firestore = require('./firebase/firestore');